deleted all the unnecessary files
