+++ .settings/org.eclipse.core.resources.prefs
@@ -5,9 +5,9 @@ eclipse.preferences.version=1
 encoding//minicommunity/__init__.py=utf-8
 encoding//minicommunity/controller/__init__.py=utf-8
 encoding//minicommunity/controller/login.py=utf-8
-encoding//minicommunity/controller/loginpola.py=utf-8
 encoding//minicommunity/controller/register_member.py=utf-8
 encoding//minicommunity/minicommunity_blueprint.py=utf-8
+encoding//minicommunity/minicommunity_cache_session.py=utf-8
 encoding//minicommunity/minicommunity_config.py=utf-8
 encoding//minicommunity/minicommunity_database.py=utf-8
 encoding//minicommunity/minicommunity_logger.py=utf-8
@@ -17,9 +17,7 @@ encoding//minicommunity/model/anonybbs_delreq.py=utf-8
 encoding//minicommunity/model/member.py=utf-8
 encoding//minicommunity/model/moviebbs_comment.py=utf-8
 encoding//minicommunity/model/moviebbs_movieinfo.py=utf-8
-encoding//minicommunity/templates/login_test.html=UTF-8
-encoding//minicommunity/templates/regist.html=UTF-8
+encoding//minicommunity/templates/login.html=UTF-8
 encoding//minicommunity/templates/regist_sample.html=UTF-8
-encoding//minicommunity/templates/register_sample.html=UTF-8
 encoding/<project>=UTF-8
 encoding/runserver.py=utf-8
