+++ minicommunity/controller/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-__all__ = ['register_member','loginpola']
+__all__ = ['register_member','login']
