--- minicommunity/templates/login_ref.html
@@ -1,112 +1,0 @@
-{% extends "layout.html" %}
-{% block title %}Login{% endblock %}
-{% block head %}
-  {{ super() }}
-{% endblock %}
-
-
-{% block content %}
-    <div class="container-fluid" > <!-- bootstrap container -->
-      <div class="row">
-        {% from "_formhelpers.html" import render_field %}
-        <form name="frm" method="POST" action="{{ url_for('minicommunity.login') }}" accept-charset="UTF-8" onSubmit="return checkLogin(this)">
-        <div id="login-form" class="modal fade modal-login" data-keyboard="false" data-backdrop="static" style="display: show; position:absolute; overflow:auto;">
-
-            <div class="modal-head">
-              <legend>&nbsp;<i class="icon-user"></i> NSHC 미니커뮤니티에 오신 것을 환영합니다! </legend>
-              {% if error %}
-              <div class="alert alert-error">
-                <a class="close" data-dismiss="alert" href="#">×</a>{{ error }}
-              </div>
-              {% endif %}
-              {% if regist_email %}
-              <div class="alert alert-success">
-                <a class="close" data-dismiss="alert" href="#">×</a>
-                {{ regist_email }}의 가입을 환영합니다! 
-              </div>
-              {% elif update_email %}
-              <div class="alert alert-info">
-                <a class="close" data-dismiss="alert" href="#">×</a>
-                {{ update_email }}의 정보가 변경되습니다. 
-              </div>
-              {% endif %}
-           </div>
-          <div class="modal-body">
-              {{ render_field(form.email, class="input-block-level", placeholder="email") }}
-              {{ render_field(form.password, class="input-block-level", placeholder="Password") }}
-              {{ render_field(form.next_url, value= next_url) }}
-              <label class="checkbox">
-                  <input type="checkbox" name="remember" value="1" onClick="confirmSave(this)" > 이메일 기억하기
-              </label>
-          </div> <!-- modal-body -->
-          
-          <div class="modal-footer">
-              <button type="submit" name="submit" class="btn btn-info">로그인</button>
-              <a href="{{ url_for('minicommunity.register_user_form') }}" class="btn btn-success" >가입하기</a>
-          </div>          
-
-        
-        </div> <!-- login -->
-       </form>
-      </div> <!-- row -->
-    </div> <!-- container -->
-
-    <script type="text/javascript">
-      $('#login-form').modal('show');
-      $('#email').focus();
-      
-      function confirmSave(checkbox) {
-        var isRemember;
-        if (checkbox.checked) {
-          isRemember = confirm("이 PC에 로그인 정보를 저장하시겠습니까? PC방등의 공공장소에서는 개인정보가 유출될 수 있으니 주의해주십시오.");
-          
-          if (!isRemember) checkbox.checked = false;
-        }
-      }
-      
-      function setsave(name, value, expiredays) {
-        var today = new Date();
-        today.setDate( today.getDate() + expiredays );
-        document.cookie = name + "=" + escape( value ) + "; path=/; expires=" + today.toGMTString() + ";"
-      }
-      
-      function saveLogin(email){
-        if (email != "") {
-          setsave("email", email, 7);
-         } else {
-           setsave("email", email, -1);
-         }
-      }
-      
-      function getLogin() {
-        var cookie = document.cookie + ";";
-        var index = cookie.indexOf("username", 0);
-        var val = "";
-        if (index != -1) {
-          cookie = cookie.substring(index, cookie.length);
-          begin = cookie.indexOf("=", 0) + 1;
-          end = cookie.indexOf(";", begin);
-          val = unescape(cookie.substring(begin, end));
-         }
-        
-        if (val!= "") {
-          document.frm.email.value = val;
-          document.frm.remember.checked = true;
-        }
-      
-      }
-      
-      function checkLogin() {
-    	if (document.frm.remember.checked) {
-    	  saveLogin(document.frm.username.value);
-    	}
-    	else {    	  
-    	  saveLogin("");
-    	}
-      }
-      
-      window.onload = getLogin()
-    </script>
-
-    
-{% endblock %}
