--- minicommunity/templates/register_sample.html
@@ -1,167 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-
-		<meta http-equiv="Content-Type" content="text/html; charset=utf-8"/>
-
-		<!-- Always force latest IE rendering engine (even in intranet) & Chrome Frame
-		Remove this if you use the .htaccess -->
-		<meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
-
-		<title>aaa</title>
-		<meta name="description" content="" />
-
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		
-				<!-- Replace favicon.ico & apple-touch-icon.png in the root of your domain and delete these references -->
-		<link rel="shortcut icon" href="/favicon.ico" />
-		<link rel="apple-touch-icon" href="/apple-touch-icon.png" />
-
-		<link rel="stylesheet" href="{{ url_for('static',filename='css/bootstrap.min.css') }}" type="text/css" media="screen" title="photolog" charset="UTF-8" />
-		<link rel="stylesheet" href="{{ url_for('static',filename='css/font-awesome.min.css') }}" type="text/css" media="screen" title="photolog" charset="UTF-8" />
-        <link rel="stylesheet" href="{{ url_for('static',filename='css/bootstrap-responsive.css') }}" type="text/css" media="screen" title="photolog" charset="UTF-8" />
-        <link rel="stylesheet" href="{{ url_for('static',filename='css/bootstrap-fileupload.min.css') }}" type="text/css" media="screen" title="photolog" charset="UTF-8" />
-        <link rel="stylesheet" href="{{ url_for('static',filename='css/style.css') }}" type="text/css" media="screen" title="photolog" charset="UTF-8" />
-        <link rel="stylesheet" href="{{ url_for('static',filename='css/placeholder-error.css') }}" type="text/css" media="screen" title="photolog" charset="UTF-8" />
-		<!--[if IE 7]>
-			<link rel="stylesheet" href="{{ url_for('static',filename='css/font-awesome-ie7.min.css')}}">
-		<![endif]-->
-
-		<!--[if lt IE 9]>
-			<script src="http://html5shiv.googlecode.com/svn/trunk/html5.js"></script>
-		<![endif]-->
-
-		<script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/1.9.0/jquery.min.js"></script>
-        <script type="text/javascript" src="{{ url_for('static',filename='js/bootstrap.min.js') }}"></script>
-        <script type="text/javascript" src="{{ url_for('static',filename='js/bootstrap-fileupload.min.js') }}"></script>
-		<script type="text/javascript" src="{{ url_for('static',filename='js/bootbox.min.js') }}"></script>
-		<script type="text/javascript" src="http://maps.google.com/maps/api/js?sensor=true"></script>
-		<script type="text/javascript" src="{{ url_for('static',filename='js/gmaps.js') }}"></script>
-        <script type="text/javascript" src="{{ url_for('static',filename='js/exif.js') }}"></script>
-        <script type="text/javascript" src="{{ url_for('static',filename='js/binaryajax.js') }}"></script>
-        <script type="text/javascript" src="{{ url_for('static',filename='js/modal.js') }}"></script>
-		
-	</head>
-	<body>
-	
-<script type=text/javascript>
-  $(function() {
-    var submit_form = function(e) {
-      if ($('input[name=email]').val() != '') {
-        $.ajax({
-        	url         : '/member/check_email'
-           ,type        : 'POST'
-           ,cache       : false
-           ,data        : JSON.stringify({
-        	   email: $('input[name=email]').val()
-             })
-           ,contentType : 'application/json; charset=utf-8'
-           ,dataType    : 'json'
-           ,success     : function(data) {
-        	   email = $('input[name=email]').val();
-        	   if (data.result) {
-        	       
-                   $('#register').modal('hide');
-                   
-                   bootbox.alert(email + '는 사용 가능한 이메일입니다.', function() {
-                       $('#register').modal('show');
-                	   $('input[name=email_check]').val('Y');
-                	   $('input[name=password]').focus().select();
-                   
-                   });
-        	   } else {
-        	       $('#register').modal('hide');
-                   
-                   bootbox.alert(email + '는 이미 사용중입니다. 다른 이메일을 입력하세요.', function() {
-                       $('#regist').modal('show');
-                	   $('input[name=email]').val('');
-                       $('input[name=email]').focus().select();
-                   });
-        	   }
-             }
-           ,error: function(result) {
-               $('#register').modal('hide');
-
-        	   bootbox.alert('통신에 오류가 발생했습니다. 잠시 후에 다시 사용하세요.', function() {
-                   $('#register').modal('show');
-        	   });
-           }
-        });
-        
-      } else { 
-          $('#register').modal('hide');
-
-    	  bootbox.alert('이메일을 넣어주세요.', function() {
-              $('#register').modal('show');
-              $('input[name=email]').focus();
-    	  });
-      }
-      return false;
-    };
-        
-    $('button#email_check').bind('click', submit_form);
-    $('input[name=email]').focus();
-  });
-</script>
-	
-<!-- Button trigger modal -->
-<button type="button" class="btn btn-primary btn-lg" data-toggle="modal" data-target="#myModal">
-  회원가입하기
-</button>
-
-<!-- Modal -->
-<div class="modal fade" id="myModal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
-  <div class="modal-dialog">
-    <div class="modal-content">
-      <div class="modal-header">
-        <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-        <h4 class="modal-title" id="myModalLabel">회원가입</h4>
-      </div>
-      			<div class="modal-body">
-					<!-- 입력창 시작 -->
-					<div class="input-group">
-						<span class="input-group-addon" id="basic-addon_email">이메일</span>
-						<div class="input-group">
-							<input type="text" class="form-control" name="email" placeholder="E-mail">
-							<span class="input-group-btn">
-								<button id="email_check" class="btn btn-default" type="button">중복확인</button>
-							</span>
-						</div>
-						<!-- /input-group -->
-					</div>
-					<br/>
-					<div class="input-group">
-						<span class="input-group-addon" id="basic-addon_nickname">닉네임</span>
-						<div class="input-group">
-							<input type="text" class="form-control" placeholder="Nickname">
-							<span class="input-group-btn">
-								<button class="btn btn-default" type="button">중복확인</button>
-							</span>
-						</div>
-						<!-- /input-group -->
-					</div>
-					<br/>
-					<div class="input-group">
-						<span class="input-group-addon" id="basic-addon_password">비밀번호</span>
-						<div class="input-group">
-							<input type="password" class="form-control" placeholder="비밀번호" aria-describedby="basic-addon_password">
-						</div>
-					</div>
-					<br/>
-					<div class="input-group">
-						<span class="input-group-addon" id="basic-addon_password_confirm">비밀번호 확인</span>
-						<div class="input-group">
-							<input type="password" class="form-control" placeholder="비밀번호 확인" aria-describedby="basic-addon_password_confirm">
-						</div>
-					</div>
-					<!-- 입력창 끝 -->
-				</div>
-      <div class="modal-footer">
-        <button type="button" class="btn btn-default" data-dismiss="modal">취소</button>
-        <button type="button" class="btn btn-primary">가입</button>
-      </div>
-    </div>
-  </div>
-</div>
-</body>
-</html>
